# Speed up BQM.add_variable and BQM.add_interaction