Comparison from 0.8.1 to this commit
```
     [e8726c36]       [6e5010f2]
        44.1±20μs       32.6±0.7μs    ~0.74  bench_binary_quadratic_model.Construction.time_empty
-        24.8±4ms       12.5±0.2ms     0.50  bench_binary_quadratic_model.Construction.time_k100
-         211±6ms          112±2ms     0.53  bench_binary_quadratic_model.ConstructionVeryLarge.time_J15000
-      24.5±0.4ms       14.9±0.6ms     0.61  bench_binary_quadratic_model.Copy.time_k100
```